feat: split Material.vue
